* rainbow-mode/rainbow-mode.el: do not fail if face-property is a symbol
It turns out there are cases when `face-property' can be just a symbol and we
need to protect our selves from that, i.e. `car' should not fail. Hence,
`car-safe' is there and if it's `nil', then fall back to `face-property' as is.

See https://github.com/tarsius/hl-todo/issues/17